azure_rest_api/requirements.txt
--index-url http://dcsartifacts.dell.com/artifactory/api/pypi/python/simple/ --trusted-host dcsartifacts.dell.com
gunicorn==20.1.0
Django==4.0.7
PyJWT==2.4.0
asgiref==3.5.2
certifi==2022.6.15
cffi== 1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
msal==1.18.0
mssql-django==1.1.3
pycparser==2.21
pyodbc==4.0.34
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.8.1
whitenoise==5.3.0